labs/week-45-kafka-backpressure/services/producer/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
confluent-kafka==2.3.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
prometheus-client==0.19.0
pydantic==2.5.3
